Input Manager Script
This script needs to be added to an empty game object to work.

Right now there are booleans for WASD, up/down/left/right arrow keys. F,
left click, and right click.

E.g. leftClickDown is called when left click is pressed down once.
leftClickUp is called when left click is released once. leftClickHold is
called for the duration that the left click is held down.