Implemented Action Buttons
Don't press action zero as it will interfere with action one. Action
zero is just there as an example and actions 2 – 9 haven’t been
implemented yet, but do not break the game in any way.
I rebounded action one's confirm button from the space bar to B.
I made a player stats button that displays the current player’s stats.
It calls a public function to display the stats so we could make the
player stats display at the start of their turn with this function.