Updated Scene to contain the Music and SFX objects
To use the Music and SFX simply call them by searching for their tag.
Then get the script component off of the object.  Once that is complete
use the script to play the audio you want by using the built in
functions on SFX or by using the audio sources for Music