Tallying at the end of the game and Action bug fixes
Tallying now works correctly at the end of the game. It resets the
players' money and votes before recounting and accounts for shadow
positions.

Also fixed a bug on action buttons where the instances of the actions
were deleted while the follow of control was still in the start function
that had been deleted.