Title Screen Settings Minor Change
When the user chooses a 3X3X3 grid, it changes the slider for the number
of voters so that its max is 27 since that grid size can only hold 27
voters. If the player had a number picked that was greater than 27 it
will change the number of voters to 12 so the user doesn't accidentally
start with a board full of voters.