End Game Screens Update
I made it so all the policies display on the election victory scene.
Also the confirm button on the game over screen now takes you back to
the title scene. It should also save your settings. Let me know on the
asana sub task for the continue button under end game screens if it
doesn't.