Fixed Options in Settings Menu Forgotten
I fixed the bug with RememberSettings forgetting the settings. There was
a race condition with game controller. Rather than try to work around
the race condition, it was much simpler to deprecate RememberSettings
and go back to just using TitleScreenSettings.

I also removed some commented out code in the scripts related to
settings and in WindowGeneratorScript.