Tutorial Skeleton Update
The tutorial skeleton is finished in code. The covers may need some
adjusting still. Besides that, all that is needed is the actually text.

There is a new setting for the tutorial on the title screen. You can
only advance through the tutorial when you reach the end of a section.
The arrows adjust to only display one arrow if you are at the beginning
or the end of a section.